requirements.txt
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
python-dotenv==0.20.0
requests==2.27.1
soupsieve==2.3.2.post1
urllib3==1.26.9
pandas==1.3.3
protobuf==3.19.1
openpyxl==3.0.10
Django==3.2.16
Pillow == 9.3.0
Faker == 12.0.1
mixer == 7.2.2
django-extensions == 3.2.1
django-debug-toolbar == 3.8.1
psycopg2 == 2.9.5
django-cleanup == 7.0.0
djangorestframework == 3.14.0
markdown == 3.4.3
django-filter == 23.1